# backend/api-spec.yml
openapi: 3.0.3
info:
  title: ZombieLand API
  description: |-
    API REST pour le parc d'attractions post-apocalyptique ZombieLand.
    
    Cette API permet de gérer :
    - L'authentification et les utilisateurs (JWT)
    - Les activités et attractions du parc
    - Les catégories d'activités
    - Les réservations avec règle métier J-10
    - Les tarifs et dates d'ouverture
    - Les images des attractions
    
    **Règle métier critique** : Annulation de réservation possible uniquement si la visite est dans plus de 10 jours (sauf administrateurs).
    
    **Équipe Backend** : Jonathan Bous- Houd Alami  
  version: 1.0.0
  contact:
    name: Équipe ZombieLand
servers:
  - url: http://localhost:3001/api/v1
    description: Serveur de développement local
  - url: https://api.zombieland.com/api/v1
    description: Serveur de production
...
paths:
...
  /auth/register:
    post:
      tags:
        - Auth
      summary: Inscription d'un nouvel utilisateur
      description: Crée un compte utilisateur avec role CLIENT par défaut
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDto'
      responses:
        '201':
          description: Utilisateur créé avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '400':
          description: Données invalides ou email déjà utilisé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '500':
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
  /auth/login:
    post:
      tags:
        - Auth
      summary: Connexion utilisateur
      description: Authentification avec email et mot de passe, retourne un JWT
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDto'
      responses:
        '200':
          description: Connexion réussie
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  expires_in:
                    type: integer
                    example: 3600
                    description: Durée de vie du token en secondes
                  user:
                    $ref: '#/components/schemas/UserDto'
        '401':
          description: Email ou mot de passe incorrect
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '500':
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
  /auth/me:
    get:
      tags:
        - Auth
      summary: Récupérer le profil de l'utilisateur connecté
      description: Retourne les informations de l'utilisateur authentifié
      operationId: getProfile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Profil utilisateur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '401':
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '500':
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
    
    patch:
      tags:
        - Auth
      summary: Mettre à jour le profil de l'utilisateur connecté
      description: Modification email et/ou mot de passe
      operationId: updateProfile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileDto'
      responses:
        '200':
          description: Profil mis à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '400':
          description: Données invalides
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '401':
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '500':
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'

  # ===== USERS =====
...
  /categories:
    get:
      tags:
        - Categories
      summary: Liste de toutes les catégories
      description: Endpoint public, accessible sans authentification
      operationId: getAllCategories
      responses:
        '200':
          description: Liste des catégories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryDto'
        '500':
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
    
    post:
      tags:
        - Categories
      summary: Créer une catégorie (ADMIN)
      operationId: createCategory
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategoryDto'
      responses:
        '201':
          description: Catégorie créée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryDto'
        '400':
          description: Données invalides
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '401':
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '403':
          description: Accès interdit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '500':
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
  /categories/{id}:
    get:
      tags:
        - Categories
      summary: Détail d'une catégorie
      operationId: getCategoryById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Détail catégorie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryDto'
        '404':
          description: Catégorie non trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '500':
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
    
    patch:
      tags:
        - Categories
      summary: Modifier une catégorie (ADMIN)
      operationId: updateCategory
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCategoryDto'
      responses:
        '200':
          description: Catégorie modifiée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryDto'
        '400':
          description: Données invalides
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '401':
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '403':
          description: Accès interdit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '404':
          description: Catégorie non trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '500':
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
    
    delete:
      tags:
        - Categories
      summary: Supprimer une catégorie (ADMIN)
      operationId: deleteCategory
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Catégorie supprimée
        '401':
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '403':
          description: Accès interdit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '404':
          description: Catégorie non trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '500':
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'

  # ===== ATTRACTIONS =====
...
  /park-dates:
    get:
      tags:
        - ParkDates
      summary: Calendrier d'ouverture du parc
      description: Endpoint public avec filtres de dates
      operationId: getAllParkDates
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Liste des dates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ParkDateDto'
        '500':
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
    
    post:
      tags:
        - ParkDates
      summary: Créer/configurer une date (ADMIN)
      operationId: createParkDate
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateParkDateDto'
      responses:
        '201':
          description: Date créée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParkDateDto'
        '400':
          description: Données invalides
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '401':
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '403':
          description: Accès interdit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '500':
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
  /park-dates/{id}:
    get:
      tags:
        - ParkDates
      summary: Détail d'une date (ADMIN)
      operationId: getParkDateById
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Détail date
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParkDateDto'
        '401':
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '403':
          description: Accès interdit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '404':
          description: Date non trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '500':
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
    
    patch:
      tags:
        - ParkDates
      summary: Modifier une date (ADMIN)
      operationId: updateParkDate
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateParkDateDto'
      responses:
        '200':
          description: Date modifiée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParkDateDto'
        '400':
          description: Données invalides
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '401':
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '403':
          description: Accès interdit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '404':
          description: Date non trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '500':
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
    
    delete:
      tags:
        - ParkDates
      summary: Supprimer une date (ADMIN)
      operationId: deleteParkDate
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Date supprimée
        '401':
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '403':
          description: Accès interdit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '404':
          description: Date non trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '500':
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'

  # ===== PRICES =====
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenu via /auth/login
  schemas:
...
    Role:
      type: string
      enum:
        - ADMIN
        - CLIENT
      description: Rôle utilisateur
...
    UserDto:
      type: object
      properties:
        id:
          type: integer
          example: 1
        email:
          type: string
          format: email
          example: user@zombieland.com
        pseudo:
          type: string
          example: JeanZ
        role:
          $ref: '#/components/schemas/Role'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CategoryDto:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Attractions extrêmes
        description:
          type: string
          example: Sensations fortes garanties pour les amateurs d'adrénaline
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ParkDateDto:
      type: object
      properties:
        id:
          type: integer
          example: 1
        jour:
          type: string
          format: date
          example: "2025-12-25"
        is_open:
          type: boolean
          example: true
        open_hour:
          type: string
          format: time
          nullable: true
          example: "09:00:00"
          description: Heure d'ouverture du parc (format HH:mm:ss)
        close_hour:
          type: string
          format: time
          nullable: true
          example: "22:00:00"
          description: Heure de fermeture du parc (format HH:mm:ss)
        notes:
          type: string
          nullable: true
          example: Horaires étendus pour les fêtes
        created_at:
          type: string
          format: date-time
...
    RegisterDto:
      type: object
      required:
        - email
        - pseudo
        - password
        - confirmPassword
      properties:
        email:
          type: string
          format: email
          example: newuser@zombieland.com
        pseudo:
          type: string
          minLength: 3
          maxLength: 20
          example: JeanZ
        password:
          type: string
          format: password
          minLength: 8
          example: Password123!
        confirmPassword:
          type: string
          format: password
          example: Password123!
    LoginDto:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: user@zombieland.com
        password:
          type: string
          format: password
          example: Password123!
    UpdateProfileDto:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 8
    CreateCategoryDto:
      type: object
      required:
        - name
        - description
      properties:
        name:
          type: string
          example: Spectacles
        description:
          type: string
          example: Shows et animations en live
    UpdateCategoryDto:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
...
    CreateParkDateDto:
      type: object
      required:
        - jour
        - is_open
      properties:
        jour:
          type: string
          format: date
          example: "2025-12-25"
        is_open:
          type: boolean
          example: true
        open_hour:
          type: string
          format: time
          nullable: true
          example: "09:00:00"
          description: Heure d'ouverture (format HH:mm:ss)
        close_hour:
          type: string
          format: time
          nullable: true
          example: "22:00:00"
          description: Heure de fermeture (format HH:mm:ss)
        notes:
          type: string
          nullable: true
          example: Horaires étendus pour les fêtes
    UpdateParkDateDto:
      type: object
      properties:
        jour:
          type: string
          format: date
          example: "2025-12-25"
        is_open:
          type: boolean
          example: true
        open_hour:
          type: string
          format: time
          nullable: true
          example: "09:00:00"
          description: Heure d'ouverture (format HH:mm:ss)
        close_hour:
          type: string
          format: time
          nullable: true
          example: "22:00:00"
          description: Heure de fermeture (format HH:mm:ss)
        notes:
          type: string
          nullable: true
          example: Horaires étendus
...
    ErrorDto:
      type: object
      required:
        - statusCode
        - message
        - error
      properties:
        statusCode:
          type: integer
          example: 400
        message:
          type: string
          example: Validation failed
        error:
          type: string
          example: Bad Request
        timestamp:
          type: string
          format: date-time